Extracted out user creation
