+++ helpers/authHelper.js
@@ -1,4 +1,7 @@
 import User from "../models/user.model.js";
+import Member from "../models/member.model.js";
+import bcrypt from "bcrypt";
+import { nanoid } from "nanoid";
 import { USER_TYPES, USER_TYPE_NAMES } from "./types.js";
 
 // Checks if access token is valid
@@ -32,3 +35,25 @@ export const hasAdminPermissions = async (accessToken) => {
   const token = await checkAccessToken(accessToken);
   return token.isValidToken && token.userPermissions === USER_TYPES.ADMIN;
 };
+
+export const sendCreateUser = async (user) => {
+  // Generate 21 length tokens
+  const access_token = nanoid();
+  const refresh_token = nanoid();
+
+  // Generate salted hash for password
+  const hash = await bcrypt.hash(user.password, 10);
+
+  // Create user document
+  const newUser = new User({
+    username: user.username,
+    permissions: user.permissions,
+    member: user.member,
+    refresh_token,
+    access_token,
+    hash,
+  });
+  const data = await newUser.save();
+
+  return data;
+}
